git-svn-id: https://repo.acin.tuwien.ac.at/v4r/stable/strandsv4r@8483 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ v4r/SurfaceUtils/EPBase.cpp
@@ -82,7 +82,7 @@ void EPBase::setNormals(const pcl::PointCloud<pcl::Normal>::Ptr &_normals)
     throw std::runtime_error(error_message);
   }
   
-  if ( (_normals->height!=height) || (_normals->width!=width) )
+  if ( (_normals->height!=(unsigned int)height) || (_normals->width!=(unsigned int)width) )
   {
     char* error_message = new char[200];
     sprintf(error_message,"[%c::setNormals()]: Invalid normals (not for this point cloud).",ClassName.c_str());
